Begin work on EditItineraryActivity
Add a button to MainActivity for editing the itinerary, which
opens a new activity.

+++ app/src/main/java/edu/cwru/students/cwrumapper/user/Event.java
@@ -51,6 +51,6 @@ public class Event {
 
     public double getEventLength() { return length; }
 
-    public double getRoomNumber() { return roomNumber; }
+    public int getRoomNumber() { return roomNumber; }
 
 }
